Change quantity type from INT to DOUBLE

+++ task_2.sql
@@ -1,3 +1,4 @@
+
 -- Create database
 CREATE DATABASE IF NOT EXISTS alx_book_store;
 USE alx_book_store;
@@ -41,7 +42,7 @@ CREATE TABLE IF NOT EXISTS Order_Details (
     order_detail_id INT AUTO_INCREMENT PRIMARY KEY,
     order_id INT,
     book_id INT,
-    quantity INT,
+    quantity DOUBLE,
     subtotal DECIMAL(10,2),
     FOREIGN KEY (order_id) REFERENCES Orders(order_id),
     FOREIGN KEY (book_id) REFERENCES Books(book_id)
